backend-old/requirements.txt
fastapi
uvicorn
supabase
python-dotenv
pydantic
sqlalchemy
pytest